Validation now working on the Comments part of the admin panel

+++ src/Controller/Admin/AdminCommentsController.php
@@ -36,8 +36,18 @@ class AdminCommentsController {
      * @param Application $app Silex application
      */
     public function deleteCommentAction($id, Application $app) {
-        $app['dao.comment']->delete($id);
-        $app['session']->getFlashBag()->add('success', 'Le commentaire a été supprimé avec succès.');
+        $comment = $app['dao.comment']->find($id);
+
+        $errors = $app['validator']->validate($comment);
+        if (count($errors) > 0) {
+            foreach ($errors as $error) {
+                $app['session']->getFlashBag()->add('error', $error->getMessage());
+            }
+        }
+        else {
+            $app['dao.comment']->delete($id);
+            $app['session']->getFlashBag()->add('success', 'Le commentaire a été supprimé avec succès.');
+        }
 
         // Redirect to admin home page
         return $app->redirect($app['url_generator']->generate('admin'));
